reorganize folder structure, move handEvluation into a specific folder, so adapt the import
