+++ backend/models/utils/handsEvaluations/pokerHandEvaluatorFlush.js
@@ -1,5 +1,5 @@
-const PokerCard = require("../pokerCard");
-const Deck = require("../deck");
+const PokerCard = require("../../pokerCard");
+const Deck = require("../../deck");
 
 function hasFlush(cards) {
   const suits = new Set();
