+++ backend/models/utils/handsEvaluations/pokerHandEvaluatorStraightFlush.js
@@ -1,5 +1,5 @@
-const PokerCard = require("./../pokerCard");
-const Deck = require("./../deck");
+const PokerCard = require("./../../pokerCard");
+const Deck = require("./../../deck");
 
 function hasStraightFlush(cards) {
   const suits = new Set();
